Feature/remove disk search manager (#2919)
* removed DiskSearchManager

* fixes to removal of DiskSearchManager

* simplified sqlite code

* moved BuildMode

* fixed tests

* fixed API tests
