+++ conans/client/remover.py
@@ -5,7 +5,8 @@ from conans.util.log import logger
 from conans.model.ref import PackageReference
 from conans.paths import SYSTEM_REQS, rm_conandir
 from conans.model.ref import ConanFileReference
-from conans.search.search import filter_outdated, DiskSearchManager
+from conans.search.search import filter_outdated, search_recipes,\
+    search_packages
 
 
 class DiskRemover(object):
@@ -128,8 +129,7 @@ class ConanRemover(object):
             remote = self._registry.remote(remote)
             references = self._remote_manager.search_recipes(remote, pattern)
         else:
-            disk_search = DiskSearchManager(self._client_cache)
-            references = disk_search.search_recipes(pattern)
+            references = search_recipes(self._client_cache, pattern)
         if not references:
             self._user_io.out.warn("No package recipe matches '%s'" % str(pattern))
             return
@@ -143,7 +143,7 @@ class ConanRemover(object):
                 if remote:
                     packages = self._remote_manager.search_packages(remote, reference, packages_query)
                 else:
-                    packages = disk_search.search_packages(reference, packages_query)
+                    packages = search_packages(self._client_cache, reference, packages_query)
                 if outdated:
                     if remote:
                         recipe_hash = self._remote_manager.get_conan_manifest(reference, remote).summary_hash
