+++ conans/model/requires.py
@@ -70,7 +70,7 @@ class Requirements(OrderedDict):
     def copy(self):
         """ We need a custom copy as the normal one requires __init__ to be
         properly defined. This is not a deep-copy, in fact, requirements in the dict
-        are changed by RequireResolver, and are propagated upstream
+        are changed by RangeResolver, and are propagated upstream
         """
         result = Requirements()
         for name, req in self.items():
