+++ conans/server/rest/controllers/conan_controller.py
@@ -133,14 +133,14 @@ class ConanController(Controller):
             ignorecase = request.params.get("ignorecase", True)
             if isinstance(ignorecase, str):
                 ignorecase = False if 'false' == ignorecase.lower() else True
-            search_service = SearchService(app.authorizer, app.search_manager, auth_user)
+            search_service = SearchService(app.authorizer, app.paths, auth_user)
             references = [str(ref) for ref in search_service.search(pattern, ignorecase)]
             return {"results": references}
 
         @app.route('%s/search' % conan_route, method=["GET"])
         def search_packages(conanname, version, username, channel, auth_user):
             query = request.params.get("q", None)
-            search_service = SearchService(app.authorizer, app.search_manager, auth_user)
+            search_service = SearchService(app.authorizer, app.paths, auth_user)
             conan_reference = ConanFileReference(conanname, version, username, channel)
             info = search_service.search_packages(conan_reference, query)
             return info
