+++ conans/server/rest/server.py
@@ -14,7 +14,7 @@ class ConanServer(object):
 
     def __init__(self, run_port, credentials_manager,
                  updown_auth_manager, authorizer, authenticator,
-                 file_manager, search_manager, server_version, min_client_compatible_version,
+                 file_manager, paths, server_version, min_client_compatible_version,
                  server_capabilities):
 
         assert(isinstance(server_version, Version))
@@ -29,7 +29,7 @@ class ConanServer(object):
         self.root_app = bottle.Bottle()
         self.root_app.mount("/v1/", self.api_v1)
         self.run_port = run_port
-        self.api_v1.search_manager = search_manager
+        self.api_v1.paths = paths
         self.api_v1.authorizer = authorizer
         self.api_v1.authenticator = authenticator
         self.api_v1.file_manager = file_manager
