+++ conans/server/server_launcher.py
@@ -11,7 +11,6 @@ from conans.model.version import Version
 from conans.server.migrate import migrate_and_get_server_config
 from conans import __version__ as SERVER_VERSION
 from conans.paths import conan_expand_user, SimplePaths
-from conans.search.search import DiskSearchManager
 from conans import SERVER_CAPABILITIES
 
 
@@ -37,11 +36,10 @@ class ServerLauncher(object):
 
         file_manager = get_file_manager(server_config, updown_auth_manager=updown_auth_manager)
 
-        search_manager = DiskSearchManager(SimplePaths(server_config.disk_storage_path))
-
         server_capabilities = SERVER_CAPABILITIES
+        paths = SimplePaths(server_config.disk_storage_path)
         self.ra = ConanServer(server_config.port, credentials_manager, updown_auth_manager,
-                              authorizer, authenticator, file_manager, search_manager,
+                              authorizer, authenticator, file_manager, paths,
                               Version(SERVER_VERSION), Version(MIN_CLIENT_COMPATIBLE_VERSION),
                               server_capabilities)
