+++ conans/test/remote/rest_api_test.py
@@ -209,7 +209,7 @@ class RestApiTest(unittest.TestCase):
         # Upload a conans
         conan_reference1 = ConanFileReference.loads("MyFirstConan/1.0.0@private_user/testing")
         self._upload_recipe(conan_reference1)
-        path1 = self.server.search_manager._paths.conan(conan_reference1)
+        path1 = self.server.paths.conan(conan_reference1)
         self.assertTrue(os.path.exists(path1))
         # Remove conans and packages
         self.api.remove_conanfile(conan_reference1)
@@ -224,12 +224,12 @@ class RestApiTest(unittest.TestCase):
             # Upload an package
             package_ref = PackageReference(conan_ref, sha)
             self._upload_package(package_ref)
-            folder = self.server.search_manager._paths.package(package_ref)
+            folder = self.server.paths.package(package_ref)
             self.assertTrue(os.path.exists(folder))
             folders[sha] = folder
 
         self.api.remove_packages(conan_ref, ["1"])
-        self.assertTrue(os.path.exists(self.server.search_manager._paths.conan(conan_ref)))
+        self.assertTrue(os.path.exists(self.server.paths.conan(conan_ref)))
         self.assertFalse(os.path.exists(folders["1"]))
         self.assertTrue(os.path.exists(folders["2"]))
         self.assertTrue(os.path.exists(folders["3"]))
@@ -237,7 +237,7 @@ class RestApiTest(unittest.TestCase):
         self.assertTrue(os.path.exists(folders["5"]))
 
         self.api.remove_packages(conan_ref, ["2", "3"])
-        self.assertTrue(os.path.exists(self.server.search_manager._paths.conan(conan_ref)))
+        self.assertTrue(os.path.exists(self.server.paths.conan(conan_ref)))
         self.assertFalse(os.path.exists(folders["1"]))
         self.assertFalse(os.path.exists(folders["2"]))
         self.assertFalse(os.path.exists(folders["3"]))
@@ -245,7 +245,7 @@ class RestApiTest(unittest.TestCase):
         self.assertTrue(os.path.exists(folders["5"]))
 
         self.api.remove_packages(conan_ref, [])
-        self.assertTrue(os.path.exists(self.server.search_manager._paths.conan(conan_ref)))
+        self.assertTrue(os.path.exists(self.server.paths.conan(conan_ref)))
         for sha in ["1", "2", "3", "4", "5"]:
             self.assertFalse(os.path.exists(folders[sha]))
