+++ conans/test/server/utils/server_launcher.py
@@ -9,11 +9,10 @@ from conans.util.log import logger
 from conans.util.files import mkdir
 from conans.test.utils.test_files import temp_folder
 from conans.server.migrate import migrate_and_get_server_config
-from conans.search.search import DiskSearchManager
-from conans.paths import SimplePaths
 import time
 import shutil
 from conans import SERVER_CAPABILITIES
+from conans.paths import SimplePaths
 
 
 TESTING_REMOTE_PRIVATE_USER = "private_user"
@@ -54,7 +53,6 @@ class TestServerLauncher(object):
         self.file_manager = get_file_manager(server_config, public_url=base_url,
                                              updown_auth_manager=updown_auth_manager)
 
-        self.search_manager = DiskSearchManager(SimplePaths(server_config.disk_storage_path))
         # Prepare some test users
         if not read_permissions:
             read_permissions = server_config.read_permissions
@@ -76,9 +74,9 @@ class TestServerLauncher(object):
 
         logger.debug("Storage path: %s" % self.storage_folder)
         self.port = TestServerLauncher.port
-
+        self.paths = SimplePaths(server_config.disk_storage_path)
         self.ra = ConanServer(self.port, credentials_manager, updown_auth_manager,
-                              authorizer, authenticator, self.file_manager, self.search_manager,
+                              authorizer, authenticator, self.file_manager, self.paths,
                               server_version, min_client_compatible_version,
                               server_capabilities)
         for plugin in plugins:
